2005-03-03  Kazuki Oikawa  <kazuki@panicode.com>
	* IBindCtx.cs
	* IConnectionPoint.cs
	* IConnectionPointContainer.cs
	* IEnumConnectionPoints.cs
	* IEnumConnections.cs
	* IEnumMoniker.cs
	* IEnumString.cs
	* IEnumVARIANT.cs
	* IMoniker.cs
	* IPersistFile.cs
	* IRunningObjectTable.cs
	* IStream.cs
	* ITypeComp.cs
	* ITypeInfo.cs
	* ITypeInfo2.cs
	* ITypeLib.cs
	* ITypeLib2.cs
	* EXCEPINFO.cs
	* IDLDESC.cs
	* TYPELIBATTR.cs
	* VARDESC.cs
	* INVOKEKIND.cs
	* SYSKIND.cs: Added missing attributes and corrected wrong declaration.

svn path=/trunk/mcs/; revision=41394

+++ mcs/class/corlib/System.Runtime.InteropServices.ComTypes/IDLDESC.cs
@@ -35,7 +35,7 @@ namespace System.Runtime.InteropServices.ComTypes
 	[StructLayout (LayoutKind.Sequential, CharSet = CharSet.Unicode)]
 	public struct IDLDESC
 	{
-		public int dwReserved;
+		public IntPtr dwReserved;
 		public IDLFLAG wIDLFlags;
 	}
 }
